requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
transformers>=4.30.0
requests>=2.28.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
nltk>=3.7
streamlit>=1.22.0
tqdm>=4.64.0